renamed pages without suffix page as will not be using components, instead using full page to render seperately in the ui
