+++ pages/Albums.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Albums = () => {
+  return (
+    <div>Albums</div>
+  )
+}
+
+export default Albums
