--- pages/AlbumsPage.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const AlbumsPage = () => {
-  return (
-    <div>AlbumsPage</div>
-  )
-}
-
-export default AlbumsPage
