+++ pages/Artists.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Artists = () => {
+  return (
+    <div>Artists</div>
+  )
+}
+
+export default Artists
