--- pages/ArtistsPage.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const ArtistsPage = () => {
-  return (
-    <div>ArtistsPage</div>
-  )
-}
-
-export default ArtistsPage
