--- pages/ExplorePage.jsx
@@ -1,7 +1,0 @@
-import Explore from "../components/Explore/Explore";
-
-const ExplorePage = () => {
-  return <Explore />;
-};
-
-export default ExplorePage;
