+++ pages/Genres.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Genres = () => {
+  return (
+    <div>Genres</div>
+  )
+}
+
+export default Genres
