--- pages/GenresPage.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const GenresPage = () => {
-  return (
-    <div>GenresPage</div>
-  )
-}
-
-export default GenresPage
