+++ pages/Local.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Local = () => {
+  return (
+    <div>Local</div>
+  )
+}
+
+export default Local
