--- pages/LocalPage.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const LocalPage = () => {
-  return (
-    <div>LocalPage</div>
-  )
-}
-
-export default LocalPage
