+++ pages/Radio.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Radio = () => {
+  return (
+    <div>Radio</div>
+  )
+}
+
+export default Radio
