--- pages/RadioPage.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const RadioPage = () => {
-  return (
-    <div>RadioPage</div>
-  )
-}
-
-export default RadioPage
