+++ pages/Recent.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Recent = () => {
+  return (
+    <div>Recent</div>
+  )
+}
+
+export default Recent
