--- pages/RecentPage.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const RecentPage = () => {
-  return (
-    <div>RecentPage</div>
-  )
-}
-
-export default RecentPage
